[Refactor] Extract defineNetworkSchema
